+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#89] Alternative FQDNs will be added as SANs when creating the self-signed certificate. 
 
 ## [v2.1.0] - 2025-07-16
 ### Added